MDEV-23456 fil_space_crypt_t::write_page0() is accessing an uninitialized page
buf_page_create() is invoked when page is initialized. So that
previous contents of the page ignored. In few cases, it does
buf_page_get_gen() is called to fetch the page from buffer pool.
It should take x-latch on the page. For compressed page, use the
existing free block from LRU list to create new page.

fseg_create(), fseg_create_general(): Introduce block as a parameter
where segment header is placed. It is used to avoid repetitive
x-latch on the same page.

- Change the assert to check whether the page has SX latch and X
latch in fsp0fsp.cc file